pyproject.toml
[tool.poetry]
name = "repl-nix-textpastepro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.38.0"
pytesseract = "^0.3.13"
nltk = "^3.9.1"
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"
pillow = "^10.4.0"
pymupdf = "^1.24.10"
scikit-learn = "^1.5.1"
pandas = "^2.2.2"
numpy = "^2.1.1"
jieba = "^0.42.1"
pdfplumber = "^0.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
